requirements.txt
streamlit
pandas
psycopg2-binary
sqlalchemy
google-generativeai
python-dotenv